(Working) Fixed enemies and Others
The enemies stopping was caused by a difference in where they were being drawn and there actual world position, this is as there was a different origin on the texture. I fixed this by adding an offset however this now makes all 144 pixel sprites using the code to have their texture's midpoint be the world pos. Adding a bit of confusion in collision code later down the line.